renamed nrsecureagent to nrsecurityagent, applied latest k2 changes
