+++ v3/newrelic/internal_response_writer.go
@@ -30,7 +30,7 @@ func (rw *replacementResponseWriter) Write(b []byte) (n int, err error) {
 
 	headersJustWritten(rw.thd, http.StatusOK, hdr)
 
-	secureAgent.SendEvent("INBOUND_WRITE", string(b))
+	secureAgent.SendEvent("INBOUND_WRITE", string(b), hdr)
 	return
 }
